Move Footer, add BodyClass to userpage Layout
