+++ app/src/userpages/components/Layout/layout.pcss
@@ -9,7 +9,7 @@
 }
 
 .content {
-  background-color: #EFEFEF;
+  background-color: #F8F8F8;
   padding: 2.5rem 0;
   flex-grow: 1;
 }
@@ -18,4 +18,8 @@
   .content {
     padding: 0;
   }
+}
+
+:global(body.core) {
+  overflow-y: scroll;
 }
